ovl: Allow ovl_open_realfile() to open metacopy inode
Most of the time we need to open inode containing data (and not metacopy)
but when fsync happens, in that case we need to make sure upper metacopy
inode is fsynced too. In that case we need to open metacopy inode
temporarily.

Add a parameter to ovl_open_realfile() which specifies whether to open
data inode or metacopy inode. Later fsync patch will make use of this
functionality. 

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>