fuse: Reschedule dmap worker if we are not making progress
I tested dmap freeing logic with one dmap and once in a while I will
deadlock. What was happing that a request went to virtiofsd and that
range can't be freed yet (inode lock is still held) and worker keeps
on trying to free something. It does not break out of busy loop. That
also means when pending request completes, worker for it completion
on virt queue does not run and a deadlock happens.

So if we are not making progress freeing dmap, take a break and reschedule.
That way other pending workers will get a chance to run and then
this worker can succeed.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>