FUSE: Add the remote fsnotify support capability to FUSE
Add the remote fsnotify support capability to FUSE init flags, which is
supported by the kernel only when the "CONFIG_FANOTIFY" or the
"CONFIG_INOTIFY_USER" config options are enabled in the guest kernel.
This means that if the kernel supports either fanotify or inotify then
it can also support the remote fsnotify.

If virtiofsd wants the remote fsnotify support feature enabled only then
the guest kernel will enable it. However, this means that the kernel
will suppress the local fsnotify events related to inodes within the
directory exported through virtiofs. The suppression of local events
prevents the guest from receiving duplicate events; one from the guest
kernel and one from virtiofsd.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>