Fsnotify: Add a wrapper around the fsnotify function
Generally, inotify/fanotify events are generated locally by calling the
"fsnotify" function in fs/notify/fsnotify.c and various helper functions.
However, now we expect events to arrive from the FUSE server. Thus,
without any intervention a user space application will receive two events.
One event is generated locally and one arrives from the server.

Hence, to avoid duplicate events we need to "suppress" the local events
generated by the guest kernel for FUSE inodes. To achieve this we add a
wrapper around the "fsnotify" function in fs/notify/fsnotify.c that
checks if the remote fsnotify is enabled and based on the check either it
"suppresses" or lets through a local event.

The wrapper will be called in the place of the original "fsnotify" call
that is responsible for the event notification (now renamed as
"__fsnotify").

When the remote fsnotify is not enabled, all local events will be let
through as expected. This process is completely transparent to user space.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>