FUSE,Fsnotify: Modify the fsnotify_destroy_marks function
Add the "fuse_fsnotify_update_mark" function in the fsnotify
subsystem to remove remote watches.

Modify the "fsnotify_destroy_marks" function in fs/notify/mark.c to
add support for the remote deletion of watches for FUSE inodes. Also
modify "fsnotify_put_mark" for the same reason. These functions are
specifically picked, since the kernel can also clear watches on its own
(e.g., when a process that monitored a file for an event exited)

Thus we send the fsnotify request for the deletion of the remote watch at
the lowest level within "fsnotify_destroy_marks" to catch the kernel watch
removal case.

The "fuse_fsnotify_update_mark" function in turn calls the
"fuse_fsnotify_send_request" function, to send an fsnotify request to the
FUSE server related to an inode watch.

Signed-off-by: Ioannis Angelakopoulos <iangelak@redhat.com>